Cleanup implementation
Implement probability monad in terms of the free monad. The primitive
generators are parameterised by a typeclass sampleable. `Random` is
provided as the default instance but the user can substitute in other
instances.

Next step is to add conditioning, and then rewrite existing inference
algorithms in terms of this new representation.