Add conditioning
Adding conditioning finishes the model. This model differs slightly from
the one in PPPM but is, I believe, better factored. No exceptions now
when sampling from models that have been conditioned---this is
statically prevented.

Next step is to implement the posterior inference algorithms and thereby
show that this model is sufficiently expressive.